requirements.txt
flask
faker
pandas
geopandas
pyodbc
python-dotenv
dash
SQLAlchemy